Additional files

+++ discovery-data/2.1/wddata-backup-restore.sh
@@ -7,6 +7,7 @@ typeset -r ROOT_DIR_WDDATA
 
 # shellcheck source=lib/restore-utilites.bash
 source "${ROOT_DIR_WDDATA}/lib/restore-updates.bash"
+source "${ROOT_DIR_WDDATA}/lib/function.bash"
 
 KUBECTL_ARGS=""
 WDDATA_BACKUP="wddata.tar.gz"
@@ -34,14 +35,22 @@ done
 
 echo "WDData: "
 echo "Release name: $RELEASE_NAME"
-GATEWAY_POD=`kubectl get pods ${KUBECTL_ARGS}| grep "${RELEASE_NAME}-watson-discovery-gateway" | grep -v watson-discovery-gateway-test | cut -d ' ' -f 1 | sed -n 1p`
+GATEWAY_POD=`kubectl get pods ${KUBECTL_ARGS} -o jsonpath="{.items[0].metadata.name}" -l release=${RELEASE_NAME},run=gateway`
+ORG_KUBECTL_ARGS=${KUBECTL_ARGS} 
+
+# If the number of gateway's container is greater than 1, this is WD 2.1.2 or later. Then, we use management pod to get backup/restore.
+if [ `kubectl get pods ${KUBECTL_ARGS} -o jsonpath="{.items[0].spec.containers[*].name}" -l release=${RELEASE_NAME},run=gateway | wc -w` -gt 1 ] ; then
+  export KUBECTL_ARGS="${KUBECTL_ARGS} -c management"
+fi
 
 # backup wddata
 if [ ${COMMAND} = 'backup' ] ; then
   BACKUP_FILE=${BACKUP_FILE:-"wddata_`date "+%Y%m%d_%H%M%S"`.backup"}
   echo "Start backup wddata..."
-  kubectl exec ${GATEWAY_POD} ${KUBECTL_ARGS} --  bash -c 'if [ `ls mnt | wc -l | xargs` != "0" ] ; then tar zcf /tmp/'${WDDATA_BACKUP}' wexdata/* mnt/* --exclude ".nfs*" ; else tar zcf /tmp/'${WDDATA_BACKUP}' wexdata/* ; fi'
-  kubectl cp "${GATEWAY_POD}:/tmp/${WDDATA_BACKUP}" "${BACKUP_FILE}" ${KUBECTL_ARGS}
+  kubectl exec ${GATEWAY_POD} ${KUBECTL_ARGS} --  bash -c 'rm -f /tmp/'${WDDATA_BACKUP}' && \
+  if [ `ls mnt | wc -l | xargs` != "0" ] ; then tar zcf /tmp/'${WDDATA_BACKUP}' --exclude ".nfs*" --exclude wexdata/logs wexdata/* mnt/* ; else tar zcf /tmp/'${WDDATA_BACKUP}' --exclude ".nfs*" --exclude wexdata/logs wexdata/* ; fi; code=$?; if [ $code -ne 0 -a $code -ne 1 ] ; then echo "Fatal Error"; exit $code; fi'
+  wait_cmd ${GATEWAY_POD} "tar zcf" ${KUBECTL_ARGS}
+  kube_cp_to_local ${GATEWAY_POD} "${BACKUP_FILE}" "/tmp/${WDDATA_BACKUP}" ${KUBECTL_ARGS}
   kubectl exec ${GATEWAY_POD} ${KUBECTL_ARGS} --  bash -c "rm -f /tmp/${WDDATA_BACKUP}"
   echo "Done: ${BACKUP_FILE}"
 fi
@@ -58,12 +67,15 @@ if [ ${COMMAND} = 'restore' ] ; then
     exit 1
   fi
   echo "Start restore wddata: ${BACKUP_FILE}"
-  kubectl cp "${BACKUP_FILE}" "${GATEWAY_POD}:/tmp/${WDDATA_BACKUP}" ${KUBECTL_ARGS}
-  kubectl exec ${GATEWAY_POD} ${KUBECTL_ARGS} -- bash -c "tar xf /tmp/${WDDATA_BACKUP} ; rm -f /tmp/${WDDATA_BACKUP}"
+  kube_cp_from_local ${GATEWAY_POD} "${BACKUP_FILE}" "/tmp/${WDDATA_BACKUP}" ${KUBECTL_ARGS}
+  kubectl exec ${GATEWAY_POD} ${KUBECTL_ARGS} -- bash -c "tar xf /tmp/${WDDATA_BACKUP} --exclude *.lck ; rm -f /tmp/${WDDATA_BACKUP}"
+  wait_cmd ${GATEWAY_POD} "tar xf" ${KUBECTL_ARGS}
   echo "Restore Done"
   echo "Applying updates"
-  ./lib/restore-updates.bash
+  . ./lib/restore-updates.bash
   wddata_updates
   echo "Completed Updates"
   echo
 fi
+
+export KUBECTL_ARGS=${ORG_KUBECTL_ARGS}
